Laser and "radio" added
added a laser button where the user can press a button and they can
shoot lazers, also added 2 songs in which the player can play an arcade
song or a calm song of their choice and they can also stop the song if
they dont want any. Before it was the arcade song playing in the
background.